requirements.txt
flask>=3.1.0
SQLAlchemy>=2.0.0
Flask-Session>=0.8.0
flask-sqlalchemy>=3.1.0
requests>=2.32.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pymupdf>=1.24.0
python-docx>=1.1.0
openpyxl>=3.11.0
reportlab>=4.2.0
pytest>=8.0.0
anthropic>=0.25.0